feat: reorganize enums into separate package and add VoteOption enum
