--- src/main/java/github/io/api_voting_challenge/model/Role.java
@@ -1,6 +1,0 @@
-package github.io.api_voting_challenge.model;
-
-public enum Role {
-    ADMIN,
-    USER
-}
