+++ src/main/java/github/io/api_voting_challenge/model/enums/Role.java
@@ -1,0 +1,6 @@
+package github.io.api_voting_challenge.model.enums;
+
+public enum Role {
+    ADMIN,
+    USER
+}
