+++ src/main/java/github/io/api_voting_challenge/model/enums/Status.java
@@ -1,4 +1,4 @@
-package github.io.api_voting_challenge.model;
+package github.io.api_voting_challenge.model.enums;
 
 public enum Status {
     PENDING,
